Solution to Input Two Floats and Print Their Product

+++ Input Two Floats and Print Their Product@@6788d5012db25efec946f93e/code.c
@@ -2,7 +2,7 @@
 
 
 
-int main() {
+float main() {
     float a,b;
     float prod;
     scanf("%f %f",&a,&b);
